+++ src/OpenTelemetry.Instrumentation.AspNet/AspNetTraceInstrumentationOptions.cs
@@ -83,7 +83,7 @@ public class AspNetTraceInstrumentationOptions
     /// Gets or sets a value indicating whether the exception will be recorded as ActivityEvent or not.
     /// </summary>
     /// <remarks>
-    /// See: <see href="https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/semantic_conventions/exceptions.md"/>.
+    /// See: <see href="https://github.com/open-telemetry/semantic-conventions/blob/main/docs/exceptions/exceptions-spans.md"/>.
     /// </remarks>
     public bool RecordException { get; set; }